+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.9.0] - 2023-06-21
+
+### Changed
+- speculos: Default touch duration changed from 0.5sec to 0.1sec
+- speculos: Control the ticker allowing to avoid any race issue when comparing screen and accessing OCR result
+- speculos: Bump minimal speculos version to 0.2.5
 
 ## [1.8.2] - 2023-05-31